move module helper to cosmos
